[Server] Update

+++ [Server]BookOnline/routes/Gerne.js
@@ -6,7 +6,10 @@ var myParser = require("body-parser");
 var router = express();
 var jsonParser = myParser.json();
 var pool = require('../db');
+var verify = require('./VerifyToken');
+var admin = false;
 
+//them the loai moi
 function add_gerne(req,res) {
 
     pool.getConnection(function(err,connection){
@@ -20,7 +23,7 @@ function add_gerne(req,res) {
         connection.query("insert into theloai set TenTheLoai = ?",req.body.TenTheLoai,function(err,rows){
             connection.release();
             if(!err) {
-                res.json(rows);
+                res.json({success:true, message:"Added new gerne!"});
             }
         });
         console.log(req.body.TenTheLoai);
@@ -32,57 +35,104 @@ function add_gerne(req,res) {
     });
 }
 
+//tim the loai trong danh sach
+function getGerne(IdTheLoai, done){
+    pool.getConnection(function(err,connection){
+        if (err) {
+            res.json({"code" : 100, "status" : "Error in connection database"});
+            return;
+        }
+
+        console.log('connected as id ' + connection.threadId);
+
+        connection.query('SELECT * FROM theloai WHERE IdTheLoai = ? LIMIT 1', [IdTheLoai], function(err, rows, fields) {
+            if (err) throw err;
+            done(rows[0]);
+        });
+
+        connection.on('error', function(err) {
+            res.json({"code" : 100, "status" : "Error in connection database"});
+            return;
+        });
+    });
+}
+
+//cap nhat the loai trong danh sach
 function update_gerne(req,res) {
 
-    pool.getConnection(function(err,connection){
-        if (err) {
-            res.json({"code" : 100, "status" : "Error in connection database"});
-            return;
+    if (!req.body.IdTheLoai){
+        return res.json({success: false, message: 'You must send IdTheLoai'});
+    }
+
+    getGerne(req.body.IdTheLoai, function(gerne){
+        if(!gerne){
+            return res.json({success: false, message: 'Updated failed. Not found this data in database.'});
         }
+        else{
+            pool.getConnection(function(err,connection){
+                if (err) {
+                    res.json({"code" : 100, "status" : "Error in connection database"});
+                    return;
+                }
 
-        console.log('connected as id ' + connection.threadId);
+                console.log('connected as id ' + connection.threadId);
 
-        var post = {TenTheLoai: req.body.TenTheLoai};
-        connection.query("update theloai set ? where IdTheLoai = ?",[post, req.body.IdTheLoai],function(err,rows){
-            connection.release();
-            if(!err) {
-                res.json(rows);
-            }
-        });
-        //console.log(req.body.TenTheLoai);
+                var post = {TenTheLoai: req.body.TenTheLoai};
+                connection.query("update theloai set ? where IdTheLoai = ?",[post, req.body.IdTheLoai],function(err,rows){
+                    connection.release();
+                    if(!err) {
+                        res.json({success:true, message:"Updated successfully!"});
+                    }
+                });
+                //console.log(req.body.TenTheLoai);
 
-        connection.on('error', function(err) {
-            res.json({"code" : 100, "status" : "Error in connection database"});
-            return;
-        });
+                connection.on('error', function(err) {
+                    res.json({"code" : 100, "status" : "Error in connection database"});
+                    return;
+                });
+            });
+        }
     });
 }
 
+//xoa the loai
 function delete_gerne(req,res) {
 
-    pool.getConnection(function(err,connection){
-        if (err) {
-            res.json({"code" : 100, "status" : "Error in connection database"});
-            return;
+    if (!req.body.IdTheLoai){
+        return res.json({success: false, message: 'You must send IdTheLoai'});
+    }
+
+    getGerne(req.body.IdTheLoai, function(gerne){
+        if(!gerne){
+            return res.json({success: false, message: 'Updated failed. Not found this data in database.'});
         }
+        else{
+            pool.getConnection(function(err,connection){
+                if (err) {
+                    res.json({"code" : 100, "status" : "Error in connection database"});
+                    return;
+                }
 
-        console.log('connected as id ' + connection.threadId);
+                console.log('connected as id ' + connection.threadId);
 
-        connection.query("delete from theloai where IdTheLoai = ? ",req.body.IdTheLoai,function(err,rows){
-            connection.release();
-            if(!err) {
-                res.json(rows);
-            }
-        });
-        //console.log(req.body.TenTheLoai);
+                connection.query("delete from theloai where IdTheLoai = ? ",req.body.IdTheLoai,function(err,rows){
+                    connection.release();
+                    if(!err) {
+                        res.json({success:true, message: "Deleted successfully!"});
+                    }
+                });
+                //console.log(req.body.TenTheLoai);
 
-        connection.on('error', function(err) {
-            res.json({"code" : 100, "status" : "Error in connection database"});
-            return;
-        });
+                connection.on('error', function(err) {
+                    res.json({"code" : 100, "status" : "Error in connection database"});
+                    return;
+                });
+            });
+        }
     });
 }
 
+//xem tat ca the loai
 function get_all_gerne(req,res) {
 
     pool.getConnection(function(err,connection){
@@ -93,7 +143,7 @@ function get_all_gerne(req,res) {
 
         console.log('connected as id ' + connection.threadId);
 
-        connection.query("SELECT IdTheLoai, TenTheLoai FROM theloai",function(err,rows){
+        connection.query("SELECT * FROM theloai",function(err,rows){
             connection.release();
             if(!err) {
                 //setValue(rows);
@@ -108,18 +158,37 @@ function get_all_gerne(req,res) {
     });
 }
 
-router.post("/new",jsonParser, function(req,res){
-    add_gerne(req,res);
+//Authentication
+router.use("/api", function(req, res, next){
+    verify.verifyToken(req, res, next);
+    admin = verify.verifyAdmin(req, res);
 });
 
-router.put("/update",jsonParser, function(req,res){
-    update_gerne(req,res);
+//only admin
+router.post("/api/new",jsonParser, function(req,res){
+    if(admin)
+        add_gerne(req,res);
+    else
+        res.json({success:false, message:"You are not allowed to access this site"});
 });
 
-router.delete("/delete",jsonParser, function(req,res){
-    delete_gerne(req,res);
+//only admin
+router.put("/api/update",jsonParser, function(req,res){
+    if(admin)
+        update_gerne(req,res);
+    else
+        res.json({success:false, message:"You are not allowed to access this site"});
 });
 
+//only admin
+router.delete("/api/delete",jsonParser, function(req,res){
+    if(admin)
+        delete_gerne(req,res);
+    else
+        res.json({success:false, message:"You are not allowed to access this site"});
+});
+
+//public
 router.get("/get/all",function(req,res){
     get_all_gerne(req,res);
 });
